requirements.txt
Automat==24.8.1
Brotli==1.0.9
Django==5.1.2
Pillow==9.3.0
Protego==0.3.1
PyDispatcher==2.0.7
PySocks==1.7.1
Scrapy==2.11.2
Shapely==1.8.5.post1
Werkzeug==2.0.3
aiohttp==3.8.4
aiosignal==1.3.1
ansicon==1.89.0
anyio==4.6.2.post1
asgiref==3.8.1
async-generator==1.10
async-timeout==4.0.2
attrs==24.2.0
base58==2.1.1
beautifulsoup4==4.9.3
bitcoin==1.1.42
blessed==1.20.0
blinker==1.5
boto3==1.24.90
botocore==1.27.90
certifi==2022.6.15.1
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
chromedriver-autoinstaller==0.3.1
click==8.0.4
cligj==0.7.2
colorama==0.4.6
constantly==23.10.4
contourpy==1.0.6
cryptography==39.0.2
cssselect==1.2.0
cycler==0.11.0
defusedxml==0.7.1
django-googlesearch==0.2
dnspython==2.7.0
ecdsa==0.18.0
editor==1.6.6
email-finder-website==0.4
emailfinder==0.3.0b0
emailfinderpy==1.0.1
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fake-useragent==1.5.1
filelock==3.16.1
fonttools==4.38.0
frozenlist==1.3.3
geojson==2.5.0
geojson-precision==0.0.1
googlesearch-python-extended==2020.1.0
googlesearch-tool==1.0.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inquirer==3.4.0
itemadapter==0.9.0
itemloaders==1.3.2
jinxed==1.3.0
jmespath==1.0.1
jsonschema==3.2.0
jsonseq==1.0.0
kaitaistruct==0.10
kiwisolver==1.4.4
lxml==5.3.0
mapbox-tilesets==1.8.1
matplotlib==3.6.2
mercantile==1.1.6
multidict==6.0.4
numpy==1.24.1
openai==0.27.2
openpyxl==3.1.5
outcome==1.2.0
packaging==22.0
pandas==1.4.3
pandas-stubs==1.2.0.53
parsel==1.9.1
pip==24.2
prompt-toolkit==3.0.48
protobuf==4.21.12
py3-validate-email==1.0.9
py3dns==4.0.2
pyOpenSSL==23.0.0
pyasn1==0.4.8
pyasn1-modules==0.4.1
pycoin==0.90.20190728
pycparser==2.21
pydivert==2.1.0
pyfiglet==1.0.2
pyparsing==3.0.9
pyproj==3.4.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pyuseragents==1.0.5
queuelib==1.7.0
readchar==4.2.0
requests==2.32.3
requests-file==2.1.0
requests-futures==1.0.1
requests-toolbelt==0.10.0
runs==1.2.2
s3transfer==0.6.0
selenium==4.26.0
selenium-stealth==1.0.6
selenium-wire==4.6.4
service-identity==24.2.0
setuptools==75.1.0
sherlock-project==0.15.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.5.1
stem==1.8.2
tldextract==5.1.2
tomba-io==1.0.3
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.9.2
twisted==24.10.0
typing-extensions==4.12.2
tzdata==2024.2
urllib3==1.26.12
urllib3-secure-extra==0.1.0
validate-email-address==1
w3lib==2.2.1
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.1
wget==3.2
wheel==0.44.0
wsproto==1.2.0
xmod==1.8.1
yarl==1.8.2
zope.interface==7.1.1
zstandard==0.20.0
autocommand==2.2.2
backports.tarfile==1.2.0
importlib-metadata==8.0.0
importlib-resources==6.4.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
more-itertools==10.3.0
platformdirs==4.2.2
tomli==2.0.1
typeguard==4.3.0
zipp==3.19.2